Started a rough ui with fragments off the main activity.
fragment_register.xml is incomplete and needs a multiple selection for instrument implemented.

fragment_main.xml is the first screen asks the user to register or login.

fragment_login.xml needs a database before username/password validation can be implemented.

fragment_main.xml commented out the bottom navigation bar for now. Need to find a way to hide it until the user is logged in.

userProfile class to keep track of user objects to be put into the database after it's been implemented.

profileMap.java This java class has a few public functions to map instruments and genders to integers. I think this might make it easier to do comparisons later on, also might be necessary if we allow users to have more than one secondary instrument.

MainActivity onValidLogin and onRegisterComplete not implemented. Need to implement the interface they point to first.

BottomNavActivity - Started implementation of the activities main routine, consists of an activity with a bottom navigation bar, and 3 fragments