Implemented a basic match. When a user swipes right, the liked user is saved to an array associated with the current user.
The other users array is then checked to see if the current user is a member of the other user's liked users.

Finally on a match, each user adds an element to their conversationList array, and two conversation objects are initialized and associated with the respective ParseUsers and ConversationList objects.